+++ resources/externs/stripe.js
@@ -1,0 +1,15 @@
+var StripeCheckout = {
+  "require": function () {},
+  "__app": {
+    "trackOpen": function () {},
+    "setHost": function () {},
+    "configure": function () {},
+    "open": function () {},
+    "configurations": function () {},
+    "checkouts": function () {},
+    "host": {},
+    "totalButtons": {}
+  },
+  "open": function () {},
+  "configure": function () {}
+};